Refactor minimal API in Program.cs out to Controllers
